DEV: Use better names for initializers and files
