+++ assets/javascripts/discourse/controllers/generate-paper-key.js
@@ -1,7 +1,7 @@
 import { ajax } from "discourse/lib/ajax";
 import ModalFunctionality from "discourse/mixins/modal-functionality";
 import { getIdentity } from "discourse/plugins/discourse-encrypt/lib/discourse";
-import { generatePaperKey } from "discourse/plugins/discourse-encrypt/lib/paper_key";
+import { generatePaperKey } from "discourse/plugins/discourse-encrypt/lib/paper-key";
 import { exportIdentity } from "discourse/plugins/discourse-encrypt/lib/protocol";
 
 export default Ember.Controller.extend(ModalFunctionality, {
