+++ assets/javascripts/discourse/widgets/encrypted-post-timer-counter.js
@@ -1,8 +1,8 @@
-import { createWidget } from "discourse/widgets/widget";
-import { iconNode } from "discourse-common/lib/icon-library";
 import { later } from "@ember/runloop";
-import { h } from "virtual-dom";
+import { iconNode } from "discourse-common/lib/icon-library";
+import { createWidget } from "discourse/widgets/widget";
 import i18n from "I18n";
+import { h } from "virtual-dom";
 
 createWidget("encrypted-post-timer-counter", {
   tagName: "div.encrypted-post-timer-counter",
