+++ assets/javascripts/lib/paper-key.js
