+++ test/javascripts/lib/protocol-v0-test.js
@@ -8,7 +8,7 @@ import {
   _importPublicKey,
   _stringToBuffer,
   generateIdentity,
-} from "discourse/plugins/discourse-encrypt/lib/protocol_v0";
+} from "discourse/plugins/discourse-encrypt/lib/protocol-v0";
 import { test } from "qunit";
 
 QUnit.module("discourse-encrypt:lib:protocol_v0");
